+++ src/module/settings/RulsetSettings.ts
@@ -67,6 +67,10 @@ export default class RulesetSettings extends AdvancedSettings {
     settings.push(booleanSetting("showTimeframe", false));
     settings.push(stringChoiceSetting('showHullAndArmor', "armorOnly", true, TWODSIX.VehicleProtection));
     settings.push(stringSetting("sorcerySkill", "Sorcery", false, "world"));
+    settings.push(booleanSetting("useNationality", false));
+    settings.push(booleanSetting("animalsUseHits", false));
+    settings.push(booleanSetting("animalsUseLocations", false));
+    settings.push(booleanSetting("displayReactionMorale", false));
     return settings;
   }
 }